fix: add task validate before delete

+++ src/graphql/mutations/task/task-destroy-mutation.ts
@@ -6,6 +6,16 @@ export const TaskDestroyMutation = mutationField('taskDestroy', {
     taskId: nonNull('UUID'),
   },
   resolve: async (parent, { taskId }, ctx) => {
+    const task = await ctx.prisma.task.findUnique({
+      where: {
+        id: taskId,
+      },
+    });
+
+    if (!task) {
+      throw new Error(`Task not found`);
+    }
+
     const data = await ctx.prisma.task.delete({
       where: {
         id: taskId,
